refactor: use clear() to eagerly clean the builder slices
