+++ .circleci/config.yml
@@ -10,7 +10,7 @@ jobs:
     steps:
       - checkout
       - go/install:
-          version: 1.20.4
+          version: 1.21.0
       - run:
           command: |
             list=$(go list ./... | circleci tests split --split-by=timings)
