+++ internal/cmds/builder_put_121.go
@@ -1,0 +1,11 @@
+//go:build go1.21
+
+package cmds
+
+func Put(cs *CommandSlice) {
+	clear(cs.s)
+	cs.s = cs.s[:0]
+	cs.l = -1
+	cs.r = 0
+	pool.Put(cs)
+}
